refactor: start using PgStore everywhere
