+++ src/api/routes/bns/namespaces.ts
@@ -1,12 +1,12 @@
 import * as express from 'express';
 import { asyncHandler } from '../../async-handler';
-import { DataStore } from '../../../datastore/common';
+import { PgStore } from '../../../datastore/pg-store';
 import { parsePagingQueryInput } from '../../../api/pagination';
 import { isUnanchoredRequest } from '../../query-helpers';
 import { BnsErrors } from '../../../bns-constants';
 import { BnsGetAllNamespacesResponse } from '@stacks/stacks-blockchain-api-types';
 
-export function createBnsNamespacesRouter(db: DataStore): express.Router {
+export function createBnsNamespacesRouter(db: PgStore): express.Router {
   const router = express.Router();
 
   router.get(
