+++ src/api/routes/bns/pricing.ts
@@ -1,6 +1,5 @@
 import * as express from 'express';
 import { asyncHandler } from '../../async-handler';
-import { DataStore, DbBnsNamespace } from '../../../datastore/common';
 import {
   makeRandomPrivKey,
   getAddressFromPrivateKey,
@@ -9,9 +8,6 @@ import {
   bufferCVFromString,
   callReadOnlyFunction,
   ClarityType,
-  tupleCV,
-  uintCV,
-  listCV,
   ChainID,
 } from '@stacks/transactions';
 import { GetStacksNetwork, getBnsContractID } from './../../../bns-helpers';
@@ -20,8 +16,9 @@ import {
   BnsGetNamespacePriceResponse,
 } from '@stacks/stacks-blockchain-api-types';
 import { isValidPrincipal, logger } from './../../../helpers';
+import { PgStore } from '../../../datastore/pg-store';
 
-export function createBnsPriceRouter(db: DataStore, chainId: ChainID): express.Router {
+export function createBnsPriceRouter(db: PgStore, chainId: ChainID): express.Router {
   const router = express.Router();
   const stacksNetwork = GetStacksNetwork(chainId);
