+++ src/event-replay/event-replay.ts
@@ -1,9 +1,10 @@
 import * as path from 'path';
 import * as fs from 'fs';
-import { cycleMigrations, dangerousDropAllTables, PgDataStore } from '../datastore/postgres-store';
 import { startEventServer } from '../event-stream/event-server';
 import { getApiConfiguredChainID, httpPostRequest, logger } from '../helpers';
 import { findTsvBlockHeight, getDbBlockHeight } from './helpers';
+import { PgPrimaryStore } from '../datastore/pg-primary-store';
+import { cycleMigrations, dangerousDropAllTables } from '../datastore/migrations';
 
 enum EventImportMode {
   /**
@@ -47,7 +48,7 @@ export async function exportEventsAsTsv(
   console.log(`Export event data to file: ${resolvedFilePath}`);
   const writeStream = fs.createWriteStream(resolvedFilePath);
   console.log(`Export started...`);
-  await PgDataStore.exportRawEventRequests(writeStream);
+  await PgPrimaryStore.exportRawEventRequests(writeStream);
   console.log('Export successful.');
 }
 
@@ -83,7 +84,7 @@ export async function importEventsFromTsv(
     default:
       throw new Error(`Invalid event import mode: ${importMode}`);
   }
-  const hasData = await PgDataStore.containsAnyRawEventRequests();
+  const hasData = await PgPrimaryStore.containsAnyRawEventRequests();
   if (!wipeDb && hasData) {
     throw new Error(`Database contains existing data. Add --wipe-db to drop the existing tables.`);
   }
@@ -108,7 +109,7 @@ export async function importEventsFromTsv(
     console.log(`Ignoring all prunable events before block height: ${prunedBlockHeight}`);
   }
 
-  const db = await PgDataStore.connect({
+  const db = await PgPrimaryStore.connect({
     usageName: 'import-events',
     skipMigrations: true,
     withNotifier: false,
@@ -123,7 +124,7 @@ export async function importEventsFromTsv(
   });
 
   const readStream = fs.createReadStream(resolvedFilePath);
-  const rawEventsIterator = PgDataStore.getRawEventRequests(readStream, status => {
+  const rawEventsIterator = PgPrimaryStore.getRawEventRequests(readStream, status => {
     console.log(status);
   });
   // Set logger to only output for warnings/errors, otherwise the event replay will result
